requirements.txt
# Flask e extensões
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# MongoDB
pymongo==4.6.1
mongoengine==0.27.0

# Validação e serialização
marshmallow==3.20.2
pydantic==2.5.2

# Utilitários
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0

# Desenvolvimento
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Logging e monitoramento
structlog==23.2.0

# OpenAI (para futuras integrações de IA)
openai==1.6.1

# Gunicorn para produção
gunicorn==21.2.0
